+++ NEWS.md
@@ -1,3 +1,9 @@
+# codebook 0.5.8.9000
+- fix bugs in plot_labelled
+- fix bugs when variables are entirely missings
+- escape HTML in various labels, use safe names for anchors, figures
+- reliability functions no longer garble names
+
 # codebook 0.5.8
 - don't write files into anything but tempdir